pyproject.toml
[project]
name = "hcm-mcp-server"
version = "0.2.0"
description = "A FastAPI-based Model Context Protocol (MCP) server for Highway Capacity Manual (HCM) analysis and transportation engineering calculations."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu==1.11.0",
    "fastapi==0.116.1",
    "fastapi-mcp==0.3.6",
    "langchain==0.3.26",
    "mcp[cli]==1.11.0",
    "numpy>=2.2.4",
    "pymupdf>=1.26.1",
    "pymupdf4llm>=0.0.25",
    "sentence-transformers==5.0.0",
    "transportations_library==0.1.4",
    "python-dotenv==1.1.1",
    "chromadb==1.0.15",
]
